requirements.txt
a2a-sdk
uvicorn
beeai-framework[a2a]
crewai
beeai-framework
beeai-sdk
acp-sdk